requirements.txt
Flask
rembg[cli]
Pillow
waitress
colorgram.py